+++ LuckyoneUI/Modules/Chat/Chat.lua
@@ -163,4 +163,8 @@ function Private:Setup_Chat(installer)
 	end
 
 	Private:Print(L["Chat setup successful."])
+
+	if Private.isRetail or Private.isTBC then
+		Private:Print(L["Your chat will keep disappearing unless you move it in Blizzard's Edit Mode and save the profile there."])
+	end
 end