+++ LuckyoneUI/Modules/Tags/Classic.lua
@@ -93,12 +93,14 @@ E:AddTag('luckyone:health:percent-with-absorbs:nostatus', 'UNIT_HEALTH UNIT_MAXH
 end, Private.isClassic)
 E:AddTagInfo('luckyone:health:percent-with-absorbs:nostatus', Private.Name, L["Displays the unit's current health as a percentage with absorb values, without decimals and without status"], nil, Private.isClassic)
 
--- Display percentage power with powercolor and hides power at 0
+-- Display percentage power with powercolor
 E:AddTag('luckyone:power:percent-color', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT UNIT_DISPLAYPOWER', function(unit)
 	local min, max = UnitPower(unit), UnitPowerMax(unit)
+	if max == 0 then return end
+
 	local pType, pToken, altR, altG, altB = UnitPowerType(unit)
 	local color = _COLORS.power[pToken] or (altR and (altR > 1 or altG > 1 or altB > 1) and Private.Tags.Hex(altR / 255, altG / 255, altB / 255)) or Private.Tags.Hex(_COLORS.power[pType] or _COLORS.power.MANA)
-	local percentage = (min == 0 or max == 0) and '' or floor(min / max * 100 + .5)
+	local percentage = floor(min / max * 100 + .5)
 
 	if altR and not _COLORS.power[pToken] then
 		if altR <= 1 and altG <= 1 and altB <= 1 then
@@ -108,16 +110,20 @@ E:AddTag('luckyone:power:percent-color', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT UNIT
 		color = Private.Tags.Hex(color)
 	end
 
-	return color .. percentage
+	if percentage ~= 0 then
+		return color .. percentage
+	end
 end)
-E:AddTagInfo('luckyone:power:percent-color', Private.Name, L["Displays percentage power with powercolor and hides power at 0"])
+E:AddTagInfo('luckyone:power:percent-color', Private.Name, L["Displays percentage power without decimals with powercolor"])
 
--- Display percentage power with no color and hides power at 0
+-- Display percentage power with no color
 E:AddTag('luckyone:power:percent-nocolor', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT UNIT_DISPLAYPOWER', function(unit)
 	local min, max = UnitPower(unit), UnitPowerMax(unit)
-	return (min == 0) and '' or floor(min / max * 100 + .5)
+	if min ~= 0 and max ~= 0 then
+		return floor(min / max * 100 + .5)
+	end
 end)
-E:AddTagInfo('luckyone:power:percent-nocolor', Private.Name, L["Displays percentage power with no color and hides power at 0"])
+E:AddTagInfo('luckyone:power:percent-nocolor', Private.Name, L["Displays percentage power without decimals with no color"])
 
 -- Display percentage mana with 0 decimals
 E:AddTag('luckyone:mana:percent', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT UNIT_DISPLAYPOWER', function(unit)
@@ -146,7 +152,7 @@ E:AddTag('luckyone:level', 'UNIT_LEVEL PLAYER_LEVEL_UP', function(unit)
 end)
 E:AddTagInfo('luckyone:level', Private.Name, L["Displays the unit's level with difficultycolor if the player is not max level"])
 
--- Display mana (current) if the unit is flagged healer (Retail and Mists only)
+-- Display mana (current) if the unit is flagged healer
 E:AddTag('luckyone:healermana:current', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT UNIT_DISPLAYPOWER', function(unit)
 	local role = UnitGroupRolesAssigned(unit)
 
@@ -154,21 +160,22 @@ E:AddTag('luckyone:healermana:current', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT UNIT_
 		local color = ElvUF_colors_power.MANA
 		return Private.Tags.Hex(color.r, color.g, color.b) .. UnitPower(unit, Enum.PowerType.Mana)
 	end
-end, Private.isClassic)
-E:AddTagInfo('luckyone:healermana:current', Private.Name, L["Displays the unit's Mana with manacolor (Role: Healer)"], nil, Private.isClassic)
+end)
+E:AddTagInfo('luckyone:healermana:current', Private.Name, L["Displays the unit's Mana with manacolor (Role: Healer)"])
 
--- Display mana (percent) if the unit is flagged healer (Retail and Mists only)
+-- Display mana (percent) if the unit is flagged healer
 E:AddTag('luckyone:healermana:percent', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT UNIT_DISPLAYPOWER', function(unit)
 	local role = UnitGroupRolesAssigned(unit)
 	if role == 'HEALER' then
 		local color = ElvUF_colors_power.MANA
 		local min = UnitPower(unit, Enum.PowerType.Mana)
 		local max = UnitPowerMax(unit, Enum.PowerType.Mana)
+		if max == 0 then return end
 
 		return Private.Tags.Hex(color.r, color.g, color.b) .. E:GetFormattedText('PERCENT', min, max, 0, nil)
 	end
-end, Private.isClassic)
-E:AddTagInfo('luckyone:healermana:percent', Private.Name, L["Displays the unit's Mana with manacolor in percent (Role: Healer)"], nil, Private.isClassic)
+end)
+E:AddTagInfo('luckyone:healermana:percent', Private.Name, L["Displays the unit's Mana with manacolor in percent (Role: Healer)"])
 
 -- Displays the last (and mostly important) part of the unit's name with class color
 E:AddTag('luckyone:name:last-classcolor', 'UNIT_NAME_UPDATE UNIT_FACTION INSTANCE_ENCOUNTER_ENGAGE_UNIT', function(unit)