+++ LuckyoneUI/Profiles/ElvUI.lua
@@ -265,6 +265,10 @@ function Private:Setup_NamePlates(installer)
 	-- NamePlates CVars
 	Private:NameplateCVars()
 
+	if Private.isRetail then
+		E.db.nameplates.persistentFriendlyNP = true
+	end
+
 	-- NamePlates colors
 	E.db.nameplates.colors.auraByType = false
 	E.db.nameplates.colors.castbarDesaturate = false
@@ -834,36 +838,54 @@ function Private:Setup_ElvUI(layout)
 
 	-- Cooldown Text
 	E.db.cooldown.actionbar.colors.edge.a = 0
+	E.db.cooldown.actionbar.colors.edgeCharge.a = 0
+	E.db.cooldown.actionbar.colors.edgeLOC.a = 0
 	E.db.cooldown.actionbar.fontSize = 12
+	E.db.cooldown.actionbar.hideBling = true
 	E.db.cooldown.actionbar.offsetX = 1
 	E.db.cooldown.aurabars.fontSize = 12
+	E.db.cooldown.aurabars.hideBling = true
+	E.db.cooldown.auraindicator.reverse = true
 	E.db.cooldown.auras.colors.edge.a = 0
+	E.db.cooldown.auras.colors.swipe.a = 0
 	E.db.cooldown.auras.fontSize = 12
+	E.db.cooldown.auras.hideBling = true
 	E.db.cooldown.auras.offsetX = 1
 	E.db.cooldown.auras.offsetY = -7
 	E.db.cooldown.bags.colors.edge.a = 0
 	E.db.cooldown.bags.fontSize = 12
+	E.db.cooldown.bags.hideBling = true
 	E.db.cooldown.bags.offsetX = 1
 	E.db.cooldown.bossbutton.colors.edge.a = 0
 	E.db.cooldown.bossbutton.fontSize = 14
+	E.db.cooldown.bossbutton.hideBling = true
 	E.db.cooldown.bossbutton.offsetX = 1
 	E.db.cooldown.cdmanager.colors.edge.a = 0
 	E.db.cooldown.cdmanager.fontSize = 14
+	E.db.cooldown.cdmanager.hideBling = true
 	E.db.cooldown.cdmanager.offsetX = 1
 	E.db.cooldown.global.colors.edge.a = 0
 	E.db.cooldown.global.fontSize = 12
+	E.db.cooldown.global.hideBling = true
 	E.db.cooldown.global.offsetX = 1
 	E.db.cooldown.nameplates.colors.edge.a = 0
 	E.db.cooldown.nameplates.fontSize = 12
+	E.db.cooldown.nameplates.hideBling = true
 	E.db.cooldown.nameplates.offsetX = 1
+	E.db.cooldown.nameplates.reverse = true
 	E.db.cooldown.totemtracker.colors.edge.a = 0
 	E.db.cooldown.totemtracker.fontSize = 14
+	E.db.cooldown.totemtracker.hideBling = true
 	E.db.cooldown.totemtracker.offsetX = 1
+	E.db.cooldown.totemtracker.reverse = true
 	E.db.cooldown.unitframe.colors.edge.a = 0
 	E.db.cooldown.unitframe.fontSize = 12
+	E.db.cooldown.unitframe.hideBling = true
 	E.db.cooldown.unitframe.offsetX = 1
+	E.db.cooldown.unitframe.reverse = true
 	E.db.cooldown.zonebutton.colors.edge.a = 0
 	E.db.cooldown.zonebutton.fontSize = 14
+	E.db.cooldown.zonebutton.hideBling = true
 	E.db.cooldown.zonebutton.offsetX = 1
 
 	-- DataBars
@@ -1002,7 +1024,7 @@ function Private:Setup_ElvUI(layout)
 	-- Shared Arena
 	E.db.unitframe.units.arena.customTexts = {}
 	E.db.unitframe.units.arena.customTexts.Luckyone_HP = {
-		attachTextTo = 'Health',
+		attachTextTo = 'Frame',
 		enable = true,
 		font = Private.Font,
 		fontOutline = Private.Outline,
@@ -1013,47 +1035,47 @@ function Private:Setup_ElvUI(layout)
 		yOffset = 0
 	}
 	E.db.unitframe.units.arena.customTexts.Luckyone_Name = {
-		attachTextTo = 'Health',
+		attachTextTo = 'Frame',
 		enable = true,
 		font = Private.Font,
 		fontOutline = Private.Outline,
 		justifyH = 'RIGHT',
 		size = 12,
 		text_format = Private.isRetail and '[luckyone:name-color]' or '[luckyone:name:last-classcolor]',
-		xOffset = -3,
-		yOffset = 0
+		xOffset = -2,
+		yOffset = 14
 	}
 	E.db.unitframe.units.arena.customTexts.Luckyone_Power = {
-		attachTextTo = 'Power',
+		attachTextTo = 'Frame',
 		enable = true,
 		font = Private.Font,
 		fontOutline = Private.Outline,
-		justifyH = 'CENTER',
-		size = 11,
-		text_format = Private.isRetail and '[luckyone:power:percent-color]' or '[luckyone:power:percent-color]',
-		xOffset = 0,
-		yOffset = 0
+		justifyH = 'RIGHT',
+		size = 12,
+		text_format = '[luckyone:power:percent-color<%]',
+		xOffset = -3,
+		yOffset = -12
 	}
 
+	E.db.unitframe.units.arena.buffs.anchorPoint = 'TOPRIGHT'
 	E.db.unitframe.units.arena.buffs.countFont = Private.Font
 	E.db.unitframe.units.arena.buffs.countFontSize = 10
 	E.db.unitframe.units.arena.buffs.countPosition = 'TOPRIGHT'
-	E.db.unitframe.units.arena.buffs.countXOffset = 2
-	E.db.unitframe.units.arena.buffs.countYOffset = 0
+	E.db.unitframe.units.arena.buffs.countXOffset = 1
+	E.db.unitframe.units.arena.buffs.countYOffset = -1
 	E.db.unitframe.units.arena.buffs.growthY = 'DOWN'
 	E.db.unitframe.units.arena.buffs.maxDuration = 0
-	E.db.unitframe.units.arena.buffs.perrow = 5
+	E.db.unitframe.units.arena.buffs.numrows = 2
+	E.db.unitframe.units.arena.buffs.perrow = 4
 	E.db.unitframe.units.arena.buffs.priority = 'Whitelist,TurtleBuffs,Dispellable'
-	E.db.unitframe.units.arena.buffs.sizeOverride = 20
-	E.db.unitframe.units.arena.buffs.xOffset = -1
-	E.db.unitframe.units.arena.buffs.yOffset = -10
+	E.db.unitframe.units.arena.buffs.sizeOverride = 24
+	E.db.unitframe.units.arena.buffs.xOffset = 1
+	E.db.unitframe.units.arena.buffs.yOffset = 1
 	E.db.unitframe.units.arena.castbar.customTextFont.enable = true
 	E.db.unitframe.units.arena.castbar.customTextFont.font = Private.Font
-	E.db.unitframe.units.arena.castbar.customTextFont.fontSize = 9
 	E.db.unitframe.units.arena.castbar.customTimeFont.enable = true
 	E.db.unitframe.units.arena.castbar.customTimeFont.font = Private.Font
-	E.db.unitframe.units.arena.castbar.customTimeFont.fontSize = 9
-	E.db.unitframe.units.arena.castbar.height = 16
+	E.db.unitframe.units.arena.castbar.height = 22
 	E.db.unitframe.units.arena.castbar.iconAttachedTo = 'Castbar'
 	E.db.unitframe.units.arena.castbar.positionsGroup.yOffset = -1
 	E.db.unitframe.units.arena.castbar.strataAndLevel.useCustomLevel = true
@@ -1061,6 +1083,7 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.arena.castbar.textColor.b = 1
 	E.db.unitframe.units.arena.castbar.textColor.g = 1
 	E.db.unitframe.units.arena.castbar.textColor.r = 1
+	E.db.unitframe.units.arena.castbar.timeToHold = 2
 	E.db.unitframe.units.arena.castbar.width = 211
 	E.db.unitframe.units.arena.castbar.xOffsetText = 2
 	E.db.unitframe.units.arena.castbar.xOffsetTime = -2
@@ -1068,15 +1091,16 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.arena.debuffs.countFont = Private.Font
 	E.db.unitframe.units.arena.debuffs.countFontSize = 10
 	E.db.unitframe.units.arena.debuffs.countPosition = 'TOPRIGHT'
-	E.db.unitframe.units.arena.debuffs.countXOffset = 2
-	E.db.unitframe.units.arena.debuffs.countYOffset = 0
+	E.db.unitframe.units.arena.debuffs.countXOffset = 1
+	E.db.unitframe.units.arena.debuffs.countYOffset = -1
 	E.db.unitframe.units.arena.debuffs.desaturate = true
 	E.db.unitframe.units.arena.debuffs.growthX = 'LEFT'
 	E.db.unitframe.units.arena.debuffs.growthY = 'DOWN'
 	E.db.unitframe.units.arena.debuffs.maxDuration = 0
-	E.db.unitframe.units.arena.debuffs.perrow = 5
+	E.db.unitframe.units.arena.debuffs.numrows = 2
+	E.db.unitframe.units.arena.debuffs.perrow = 4
 	E.db.unitframe.units.arena.debuffs.priority = Private.isRetail and 'ImportantCC,blockNonPersonal,ClassDebuffs' or 'Blacklist,Personal,CCDebuffs'
-	E.db.unitframe.units.arena.debuffs.sizeOverride = 20
+	E.db.unitframe.units.arena.debuffs.sizeOverride = 24
 	E.db.unitframe.units.arena.debuffs.xOffset = -1
 	E.db.unitframe.units.arena.debuffs.yOffset = 1
 	E.db.unitframe.units.arena.disableFocusGlow = true
@@ -1086,32 +1110,30 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.arena.healPrediction.absorbStyle = 'WRAPPED'
 	E.db.unitframe.units.arena.health.attachTextTo = 'Frame'
 	E.db.unitframe.units.arena.health.text_format = ''
-	E.db.unitframe.units.arena.height = 41
+	E.db.unitframe.units.arena.height = 49
 	E.db.unitframe.units.arena.infoPanel.height = 16
 	E.db.unitframe.units.arena.middleClickFocus = true
 	E.db.unitframe.units.arena.name.attachTextTo = 'Frame'
 	E.db.unitframe.units.arena.name.text_format = ''
 	E.db.unitframe.units.arena.orientation = 'LEFT'
-	E.db.unitframe.units.arena.power.autoHide = true
-	E.db.unitframe.units.arena.power.height = 4
-	E.db.unitframe.units.arena.power.strataAndLevel.frameStrata = 'MEDIUM'
-	E.db.unitframe.units.arena.power.strataAndLevel.useCustomStrata = true
+	E.db.unitframe.units.arena.power.enable = false
 	E.db.unitframe.units.arena.power.text_format = ''
 	E.db.unitframe.units.arena.pvpclassificationindicator.size = 24
 	E.db.unitframe.units.arena.pvpclassificationindicator.xOffset = -40
 	E.db.unitframe.units.arena.pvpSpecIcon = false
-	E.db.unitframe.units.arena.pvpTrinket.size = 41
+	E.db.unitframe.units.arena.pvpTrinket.enable = false
+	E.db.unitframe.units.arena.pvpTrinket.size = 49
 	E.db.unitframe.units.arena.raidicon.attachTo = 'RIGHT'
-	E.db.unitframe.units.arena.raidicon.size = 60
-	E.db.unitframe.units.arena.raidicon.xOffset = 82
+	E.db.unitframe.units.arena.raidicon.size = 40
+	E.db.unitframe.units.arena.raidicon.xOffset = 42
 	E.db.unitframe.units.arena.raidicon.yOffset = 1
-	E.db.unitframe.units.arena.spacing = 18
+	E.db.unitframe.units.arena.spacing = 24
 	E.db.unitframe.units.arena.width = 210
 
 	-- Shared Boss
 	E.db.unitframe.units.boss.customTexts = {}
 	E.db.unitframe.units.boss.customTexts.Luckyone_HP = {
-		attachTextTo = 'Health',
+		attachTextTo = 'Frame',
 		enable = true,
 		font = Private.Font,
 		fontOutline = Private.Outline,
@@ -1122,48 +1144,48 @@ function Private:Setup_ElvUI(layout)
 		yOffset = 0
 	}
 	E.db.unitframe.units.boss.customTexts.Luckyone_Name = {
-		attachTextTo = 'Health',
+		attachTextTo = 'Frame',
 		enable = true,
 		font = Private.Font,
 		fontOutline = Private.Outline,
 		justifyH = 'RIGHT',
 		size = 12,
 		text_format = Private.isRetail and '[luckyone:name-color]' or '[luckyone:name:last-classcolor]',
-		xOffset = -3,
-		yOffset = 0
+		xOffset = -2,
+		yOffset = 14
 	}
 	E.db.unitframe.units.boss.customTexts.Luckyone_Power = {
-		attachTextTo = 'Power',
+		attachTextTo = 'Frame',
 		enable = true,
 		font = Private.Font,
 		fontOutline = Private.Outline,
-		justifyH = 'CENTER',
-		size = 11,
-		text_format = Private.isRetail and '[luckyone:power:percent-color]' or '[luckyone:power:percent-color]',
-		xOffset = 0,
-		yOffset = 0
+		justifyH = 'RIGHT',
+		size = 12,
+		text_format = '[luckyone:power:percent-color<%]',
+		xOffset = -3,
+		yOffset = -12
 	}
 
 	E.db.unitframe.units.boss.buffIndicator.enable = false
+	E.db.unitframe.units.boss.buffs.anchorPoint = 'TOPRIGHT'
 	E.db.unitframe.units.boss.buffs.countFont = Private.Font
 	E.db.unitframe.units.boss.buffs.countFontSize = 10
 	E.db.unitframe.units.boss.buffs.countPosition = 'TOPRIGHT'
-	E.db.unitframe.units.boss.buffs.countXOffset = 2
-	E.db.unitframe.units.boss.buffs.countYOffset = 0
+	E.db.unitframe.units.boss.buffs.countXOffset = 1
+	E.db.unitframe.units.boss.buffs.countYOffset = -1
 	E.db.unitframe.units.boss.buffs.growthY = 'DOWN'
-	E.db.unitframe.units.boss.buffs.perrow = 5
+	E.db.unitframe.units.boss.buffs.numrows = 2
+	E.db.unitframe.units.boss.buffs.perrow = 4
 	E.db.unitframe.units.boss.buffs.priority = 'Dispellable,RaidBuffsElvUI'
-	E.db.unitframe.units.boss.buffs.sizeOverride = 20
-	E.db.unitframe.units.boss.buffs.xOffset = -1
-	E.db.unitframe.units.boss.buffs.yOffset = -10
+	E.db.unitframe.units.boss.buffs.sizeOverride = 24
+	E.db.unitframe.units.boss.buffs.xOffset = 1
+	E.db.unitframe.units.boss.buffs.yOffset = 1
 	E.db.unitframe.units.boss.castbar.customTextFont.enable = true
 	E.db.unitframe.units.boss.castbar.customTextFont.font = Private.Font
-	E.db.unitframe.units.boss.castbar.customTextFont.fontSize = 9
 	E.db.unitframe.units.boss.castbar.customTimeFont.enable = true
 	E.db.unitframe.units.boss.castbar.customTimeFont.font = Private.Font
-	E.db.unitframe.units.boss.castbar.customTimeFont.fontSize = 9
 	E.db.unitframe.units.boss.castbar.displayTargetClass = false
-	E.db.unitframe.units.boss.castbar.height = 16
+	E.db.unitframe.units.boss.castbar.height = 22
 	E.db.unitframe.units.boss.castbar.iconAttachedTo = 'Castbar'
 	E.db.unitframe.units.boss.castbar.positionsGroup.yOffset = -1
 	E.db.unitframe.units.boss.castbar.strataAndLevel.useCustomLevel = true
@@ -1171,6 +1193,7 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.boss.castbar.textColor.b = 1
 	E.db.unitframe.units.boss.castbar.textColor.g = 1
 	E.db.unitframe.units.boss.castbar.textColor.r = 1
+	E.db.unitframe.units.boss.castbar.timeToHold = 2
 	E.db.unitframe.units.boss.castbar.width = 211
 	E.db.unitframe.units.boss.castbar.xOffsetText = 2
 	E.db.unitframe.units.boss.castbar.xOffsetTime = -2
@@ -1178,13 +1201,14 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.boss.debuffs.countFont = Private.Font
 	E.db.unitframe.units.boss.debuffs.countFontSize = 10
 	E.db.unitframe.units.boss.debuffs.countPosition = 'TOPRIGHT'
-	E.db.unitframe.units.boss.debuffs.countXOffset = 2
-	E.db.unitframe.units.boss.debuffs.countYOffset = 0
+	E.db.unitframe.units.boss.debuffs.countXOffset = 1
+	E.db.unitframe.units.boss.debuffs.countYOffset = -1
 	E.db.unitframe.units.boss.debuffs.growthX = 'LEFT'
 	E.db.unitframe.units.boss.debuffs.growthY = 'DOWN'
-	E.db.unitframe.units.boss.debuffs.perrow = 5
+	E.db.unitframe.units.boss.debuffs.numrows = 2
+	E.db.unitframe.units.boss.debuffs.perrow = 4
 	E.db.unitframe.units.boss.debuffs.priority = Private.isRetail and 'blockNonPersonal,ClassDebuffs' or 'Blacklist,Personal'
-	E.db.unitframe.units.boss.debuffs.sizeOverride = 20
+	E.db.unitframe.units.boss.debuffs.sizeOverride = 24
 	E.db.unitframe.units.boss.debuffs.xOffset = -1
 	E.db.unitframe.units.boss.debuffs.yOffset = 1
 	E.db.unitframe.units.boss.disableFocusGlow = true
@@ -1193,23 +1217,18 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.boss.fader.smooth = 0
 	E.db.unitframe.units.boss.healPrediction.absorbStyle = 'WRAPPED'
 	E.db.unitframe.units.boss.healPrediction.enable = true
-	E.db.unitframe.units.boss.health.attachTextTo = 'Frame'
 	E.db.unitframe.units.boss.health.text_format = ''
-	E.db.unitframe.units.boss.height = 41
+	E.db.unitframe.units.boss.height = 49
 	E.db.unitframe.units.boss.middleClickFocus = true
-	E.db.unitframe.units.boss.name.attachTextTo = 'Frame'
 	E.db.unitframe.units.boss.name.text_format = ''
 	E.db.unitframe.units.boss.orientation = 'LEFT'
-	E.db.unitframe.units.boss.power.autoHide = true
-	E.db.unitframe.units.boss.power.height = 4
-	E.db.unitframe.units.boss.power.strataAndLevel.frameStrata = 'MEDIUM'
-	E.db.unitframe.units.boss.power.strataAndLevel.useCustomStrata = true
+	E.db.unitframe.units.boss.power.enable = false
 	E.db.unitframe.units.boss.power.text_format = ''
 	E.db.unitframe.units.boss.raidicon.attachTo = 'RIGHT'
 	E.db.unitframe.units.boss.raidicon.size = 40
 	E.db.unitframe.units.boss.raidicon.xOffset = 42
 	E.db.unitframe.units.boss.raidicon.yOffset = 1
-	E.db.unitframe.units.boss.spacing = 18
+	E.db.unitframe.units.boss.spacing = 24
 	E.db.unitframe.units.boss.width = 210
 
 	-- Shared Focus
@@ -1232,7 +1251,7 @@ function Private:Setup_ElvUI(layout)
 		fontOutline = Private.Outline,
 		justifyH = 'CENTER',
 		size = 11,
-		text_format = Private.isRetail and '[luckyone:power:percent-color]' or '[luckyone:power:percent-color]',
+		text_format = '[luckyone:power:percent-color<%]',
 		xOffset = 0,
 		yOffset = 0
 	}
@@ -1317,7 +1336,7 @@ function Private:Setup_ElvUI(layout)
 		fontOutline = Private.Outline,
 		justifyH = 'CENTER',
 		size = 11,
-		text_format = Private.isRetail and '[luckyone:power:percent-color]' or '[luckyone:power:percent-color]',
+		text_format = '[luckyone:power:percent-color<%]',
 		xOffset = 0,
 		yOffset = 0
 	}
@@ -1370,7 +1389,7 @@ function Private:Setup_ElvUI(layout)
 		fontOutline = Private.Outline,
 		justifyH = 'CENTER',
 		size = 11,
-		text_format = '',
+		text_format = '[luckyone:power:percent-nocolor<%]',
 		xOffset = 0,
 		yOffset = 0
 	}
@@ -1436,9 +1455,6 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.player.height = 50
 	E.db.unitframe.units.player.partyIndicator.enable = false
 	E.db.unitframe.units.player.power.enable = false
-	E.db.unitframe.units.player.power.height = 4
-	E.db.unitframe.units.player.power.strataAndLevel.frameStrata = 'MEDIUM'
-	E.db.unitframe.units.player.power.strataAndLevel.useCustomStrata = true
 	E.db.unitframe.units.player.power.text_format = ''
 	E.db.unitframe.units.player.pvp.text_format = ''
 	E.db.unitframe.units.player.pvpIcon.scale = 0.85
@@ -1469,20 +1485,20 @@ function Private:Setup_ElvUI(layout)
 		fontOutline = Private.Outline,
 		justifyH = 'RIGHT',
 		size = 12,
-		text_format = Private.isRetail and '[luckyone:name-color]' or '[luckyone:name:last-classcolor][ |r» >luckyone:target:last-classcolor]',
-		xOffset = -3,
-		yOffset = 0
+		text_format = Private.isRetail and '[luckyone:name-color][ |r» >luckyone:target:name-classcolor]' or '[luckyone:name:last-classcolor][ |r» >luckyone:target:last-classcolor]',
+		xOffset = -2,
+		yOffset = 14
 	}
 	E.db.unitframe.units.target.customTexts.Luckyone_Power = {
-		attachTextTo = 'Power',
+		attachTextTo = 'Frame',
 		enable = true,
 		font = Private.Font,
 		fontOutline = Private.Outline,
-		justifyH = 'CENTER',
-		size = 11,
-		text_format = Private.isRetail and '[luckyone:power:percent-color]' or '[luckyone:power:percent-color]',
-		xOffset = 0,
-		yOffset = 0
+		justifyH = 'RIGHT',
+		size = 12,
+		text_format = '[luckyone:power:percent-color<%]',
+		xOffset = -3,
+		yOffset = -12
 	}
 
 	E.db.unitframe.units.target.aurabar.enable = false
@@ -1499,10 +1515,10 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.target.buffs.yOffset = 1
 	E.db.unitframe.units.target.castbar.customTextFont.enable = true
 	E.db.unitframe.units.target.castbar.customTextFont.font = Private.Font
-	E.db.unitframe.units.target.castbar.customTextFont.fontSize = 11
+	E.db.unitframe.units.target.castbar.customTextFont.fontSize = 12
 	E.db.unitframe.units.target.castbar.customTimeFont.enable = true
 	E.db.unitframe.units.target.castbar.customTimeFont.font = Private.Font
-	E.db.unitframe.units.target.castbar.customTimeFont.fontSize = 11
+	E.db.unitframe.units.target.castbar.customTimeFont.fontSize = 12
 	E.db.unitframe.units.target.castbar.height = 22
 	E.db.unitframe.units.target.castbar.strataAndLevel.useCustomLevel = true
 	E.db.unitframe.units.target.castbar.strataAndLevel.useCustomStrata = true
@@ -1536,7 +1552,7 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.target.orientation = 'LEFT'
 	E.db.unitframe.units.target.phaseIndicator.scale = 0.5
 	E.db.unitframe.units.target.power.autoHide = true
-	E.db.unitframe.units.target.power.detachedWidth = 260
+	E.db.unitframe.units.target.power.enable = false
 	E.db.unitframe.units.target.power.height = 4
 	E.db.unitframe.units.target.power.strataAndLevel.frameStrata = 'MEDIUM'
 	E.db.unitframe.units.target.power.strataAndLevel.useCustomStrata = true
@@ -1573,11 +1589,12 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.targettarget.fader.smooth = 0
 	E.db.unitframe.units.targettarget.healPrediction.absorbStyle = 'WRAPPED'
 	E.db.unitframe.units.targettarget.healPrediction.enable = true
-	E.db.unitframe.units.targettarget.height = 24
+	E.db.unitframe.units.targettarget.height = 22
 	E.db.unitframe.units.targettarget.name.text_format = ''
 	E.db.unitframe.units.targettarget.orientation = 'LEFT'
 	E.db.unitframe.units.targettarget.power.enable = false
 	E.db.unitframe.units.targettarget.raidicon.enable = false
+	E.db.unitframe.units.targettarget.width = 140
 
 	-- Shared Raidpet
 	E.db.unitframe.units.raidpet.buffIndicator.countFont = Private.Font
@@ -1650,27 +1667,37 @@ function Private:Setup_ElvUI(layout)
 		fontOutline = Private.Outline,
 		justifyH = 'CENTER',
 		size = 12,
-		text_format = Private.isRetail and '[luckyone:name:short-color-friendly]' or '[luckyone:name:short-classcolor]',
+		text_format = '',
 		xOffset = 0,
 		yOffset = 0
 	}
 
 	E.db.unitframe.units.party.buffs.anchorPoint = 'TOPLEFT'
 	E.db.unitframe.units.party.buffs.clickThrough = true
+	E.db.unitframe.units.party.buffs.countFont = Private.Font
+	E.db.unitframe.units.party.buffs.countFontSize = 10
+	E.db.unitframe.units.party.buffs.countPosition = 'TOPRIGHT'
+	E.db.unitframe.units.party.buffs.countXOffset = 1
+	E.db.unitframe.units.party.buffs.countYOffset = -1
 	E.db.unitframe.units.party.buffs.enable = Private.isRetail
 	E.db.unitframe.units.party.buffs.sizeOverride = 20
 	E.db.unitframe.units.party.buffs.yOffset = -21
 	E.db.unitframe.units.party.classbar.enable = false
+	E.db.unitframe.units.party.debuffs.anchorPoint = 'TOPRIGHT'
 	E.db.unitframe.units.party.debuffs.countFont = Private.Font
-	E.db.unitframe.units.party.debuffs.countXOffset = 2
+	E.db.unitframe.units.party.debuffs.countFontSize = 10
+	E.db.unitframe.units.party.debuffs.countPosition = 'TOPRIGHT'
+	E.db.unitframe.units.party.debuffs.countXOffset = 1
+	E.db.unitframe.units.party.debuffs.countYOffset = -1
 	E.db.unitframe.units.party.debuffs.enable = true
+	E.db.unitframe.units.party.debuffs.growthY = 'DOWN'
 	E.db.unitframe.units.party.debuffs.maxDuration = 0
 	E.db.unitframe.units.party.debuffs.numrows = 2
 	E.db.unitframe.units.party.debuffs.perrow = 4
 	E.db.unitframe.units.party.debuffs.priority = 'Dispellable,RaidDebuffs'
 	E.db.unitframe.units.party.debuffs.sizeOverride = 31
 	E.db.unitframe.units.party.debuffs.xOffset = 1
-	E.db.unitframe.units.party.debuffs.yOffset = (scaled and 0) or -1
+	E.db.unitframe.units.party.debuffs.yOffset = 1
 	E.db.unitframe.units.party.disableFocusGlow = true
 	E.db.unitframe.units.party.disableTargetGlow = true
 	E.db.unitframe.units.party.fader.minAlpha = 0.5
@@ -1678,7 +1705,7 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.party.healPrediction.absorbStyle = 'WRAPPED'
 	E.db.unitframe.units.party.healPrediction.enable = true
 	E.db.unitframe.units.party.health.text_format = ''
-	E.db.unitframe.units.party.height = 64
+	E.db.unitframe.units.party.height = 63
 	E.db.unitframe.units.party.horizontalSpacing = 1
 	E.db.unitframe.units.party.name.text_format = ''
 	E.db.unitframe.units.party.phaseIndicator.anchorPoint = 'LEFT'
@@ -1711,9 +1738,29 @@ function Private:Setup_ElvUI(layout)
 	-- Shared Raid1
 	E.db.unitframe.units.raid1.buffIndicator.countFont = Private.Font
 	E.db.unitframe.units.raid1.buffIndicator.countFontSize = 10
-	E.db.unitframe.units.raid1.buffIndicator.size = 6
+	E.db.unitframe.units.raid1.buffs.anchorPoint = 'TOPLEFT'
+	E.db.unitframe.units.raid1.buffs.clickThrough = true
+	E.db.unitframe.units.raid1.buffs.countFont = Private.Font
+	E.db.unitframe.units.raid1.buffs.countFontSize = 10
+	E.db.unitframe.units.raid1.buffs.countPosition = 'TOPRIGHT'
+	E.db.unitframe.units.raid1.buffs.countXOffset = 2
+	E.db.unitframe.units.raid1.buffs.countYOffset = 0
+	E.db.unitframe.units.raid1.buffs.growthY = 'DOWN'
+	E.db.unitframe.units.raid1.buffs.perrow = 5
+	E.db.unitframe.units.raid1.buffs.sizeOverride = 18
+	E.db.unitframe.units.raid1.buffs.yOffset = 1
 	E.db.unitframe.units.raid1.classbar.enable = false
+	E.db.unitframe.units.raid1.debuffs.anchorPoint = 'BOTTOMLEFT'
+	E.db.unitframe.units.raid1.debuffs.countFont = Private.Font
+	E.db.unitframe.units.raid1.debuffs.countFontSize = 10
+	E.db.unitframe.units.raid1.debuffs.countPosition = 'TOPRIGHT'
+	E.db.unitframe.units.raid1.debuffs.countXOffset = 2
+	E.db.unitframe.units.raid1.debuffs.countYOffset = 0
+	E.db.unitframe.units.raid1.debuffs.isAuraRaid = true
+	E.db.unitframe.units.raid1.debuffs.perrow = 5
 	E.db.unitframe.units.raid1.debuffs.priority = 'Dispellable,RaidDebuffs'
+	E.db.unitframe.units.raid1.debuffs.sizeOverride = 18
+	E.db.unitframe.units.raid1.debuffs.yOffset = -1
 	E.db.unitframe.units.raid1.disableFocusGlow = true
 	E.db.unitframe.units.raid1.disableTargetGlow = true
 	E.db.unitframe.units.raid1.fader.minAlpha = 0.5
@@ -1781,6 +1828,13 @@ function Private:Setup_ElvUI(layout)
 	E.db.unitframe.units.raid3.readycheckIcon.yOffset = 0
 	E.db.unitframe.units.raid3.resurrectIcon.attachTo = 'RIGHT'
 	E.db.unitframe.units.raid3.resurrectIcon.size = 18
+	E.db.unitframe.units.raid3.roleIcon.attachTo = 'Frame'
+	E.db.unitframe.units.raid3.roleIcon.damager = false
+	E.db.unitframe.units.raid3.roleIcon.enable = true
+	E.db.unitframe.units.raid3.roleIcon.position = 'LEFT'
+	E.db.unitframe.units.raid3.roleIcon.size = 12
+	E.db.unitframe.units.raid3.roleIcon.xOffset = 0
+	E.db.unitframe.units.raid3.roleIcon.yOffset = 0
 	E.db.unitframe.units.raid3.summonIcon.attachTo = 'RIGHT'
 	E.db.unitframe.units.raid3.summonIcon.size = 18
 	E.db.unitframe.units.raid3.threatStyle = 'NONE'
@@ -1791,12 +1845,12 @@ function Private:Setup_ElvUI(layout)
 	E.db.movers.AddonCompartmentMover = 'TOPRIGHT,ElvUIParent,TOPRIGHT,-3,-66'
 	E.db.movers.AlertFrameMover = 'TOP,ElvUIParent,TOP,0,-202'
 	E.db.movers.AltPowerBarMover = 'TOP,ElvUIParent,TOP,0,-22'
-	E.db.movers.ArenaHeaderMover = (scaled and 'TOPRIGHT,ElvUIParent,TOPRIGHT,-301,-240') or 'TOPRIGHT,ElvUIParent,TOPRIGHT,-322,-280'
-	E.db.movers.BagsMover = (scaled and 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-416,1') or 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-416,1' -- Check back from 1080p healing export
+	E.db.movers.ArenaHeaderMover = (scaled and 'TOPRIGHT,ElvUIParent,TOPRIGHT,-342,-240') or 'TOPRIGHT,ElvUIParent,TOPRIGHT,-402,-280'
+	E.db.movers.BagsMover = (scaled and 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-455,1') or 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-501,1'
 	E.db.movers.BelowMinimapContainerMover = 'TOPRIGHT,ElvUIParent,TOPRIGHT,-201,-193'
 	E.db.movers.BNETMover = 'TOPLEFT,ElvUIParent,TOPLEFT,1,-30'
 	E.db.movers.BossBannerMover = 'TOP,ElvUIParent,TOP,0,-202'
-	E.db.movers.BossHeaderMover = (scaled and 'TOPRIGHT,ElvUIParent,TOPRIGHT,-301,-240') or 'TOPRIGHT,ElvUIParent,TOPRIGHT,-322,-280'
+	E.db.movers.BossHeaderMover = (scaled and 'TOPRIGHT,ElvUIParent,TOPRIGHT,-342,-240') or 'TOPRIGHT,ElvUIParent,TOPRIGHT,-402,-280'
 	E.db.movers.BuffsMover = 'TOPRIGHT,ElvUIParent,TOPRIGHT,-202,-1'
 	E.db.movers.DebuffsMover = 'TOPRIGHT,ElvUIParent,TOPRIGHT,-202,-116'
 	E.db.movers.DTPanelLuckyone_ActionBars_DTMover = 'BOTTOM,ElvUIParent,BOTTOM,0,1'
@@ -1813,14 +1867,14 @@ function Private:Setup_ElvUI(layout)
 	E.db.movers.ElvAB_14 = 'TOPLEFT,ElvUIParent,TOPLEFT,1,-347'
 	E.db.movers.ElvAB_15 = 'TOPLEFT,ElvUIParent,TOPLEFT,1,-314'
 	E.db.movers.ElvUF_FocusMover = (scaled and 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-338,440') or 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-628,630'
-	E.db.movers.ElvUF_PartyMover = (scaled and 'TOPLEFT,ElvUIParent,TOPLEFT,300,-240') or 'TOPLEFT,ElvUIParent,TOPLEFT,590,-400'
+	E.db.movers.ElvUF_PartyMover = (scaled and 'TOPLEFT,ElvUIParent,TOPLEFT,322,-240') or 'TOPLEFT,ElvUIParent,TOPLEFT,610,-400'
 	E.db.movers.ElvUF_PetMover = (scaled and 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,451,360') or 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,739,500'
-	E.db.movers.ElvUF_PlayerCastbarMover = (scaled and 'BOTTOM,ElvUIParent,BOTTOM,0,411') or 'BOTTOM,ElvUIParent,BOTTOM,0,552'
+	E.db.movers.ElvUF_PlayerCastbarMover = (scaled and 'BOTTOM,ElvUIParent,BOTTOM,0,411') or 'BOTTOM,ElvUIParent,BOTTOM,0,551'
 	E.db.movers.ElvUF_PlayerMover = (scaled and 'BOTTOM,ElvUIParent,BOTTOM,-288,360') or 'BOTTOM,ElvUIParent,BOTTOM,-320,500'
 	E.db.movers.ElvUF_RaidpetMover = 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,1,458'
 	E.db.movers.ElvUF_TargetCastbarMover = (scaled and 'BOTTOM,ElvUIParent,BOTTOM,288,337') or 'BOTTOM,ElvUIParent,BOTTOM,320,477'
 	E.db.movers.ElvUF_TargetMover = (scaled and 'BOTTOM,ElvUIParent,BOTTOM,288,360') or 'BOTTOM,ElvUIParent,BOTTOM,320,500'
-	E.db.movers.ElvUF_TargetTargetMover = 'BOTTOM,ElvUIParent,BOTTOM,0,422'
+	E.db.movers.ElvUF_TargetTargetMover = 'BOTTOM,ElvUIParent,BOTTOM,380,454'
 	E.db.movers.ElvUIBagMover = (scaled and 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-1,172') or 'BOTTOMRIGHT,ElvUIParent,BOTTOMRIGHT,-1,210'
 	E.db.movers.ElvUIBankMover = (scaled and 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,1,172') or 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,1,210'
 	E.db.movers.EventToastMover = 'TOP,ElvUIParent,TOP,0,-111'
@@ -1836,6 +1890,7 @@ function Private:Setup_ElvUI(layout)
 	E.db.movers.MirrorTimer2Mover = 'TOP,ElvUIParent,TOP,0,-79'
 	E.db.movers.MirrorTimer3Mover = 'TOP,ElvUIParent,TOP,0,-98'
 	E.db.movers.ObjectiveFrameMover = 'TOPRIGHT,ElvUIParent,TOPRIGHT,-120,-230'
+	E.db.movers.PlayerPowerBarMover = (scaled and 'BOTTOM,ElvUIParent,BOTTOM,-287,341') or 'BOTTOM,ElvUIParent,BOTTOM,-320,481'
 	E.db.movers.PowerBarContainerMover = 'TOP,ElvUIParent,TOP,0,-132'
 	E.db.movers.PrivateAurasMover = (scaled and 'TOP,ElvUIParent,TOP,212,-474') or 'TOP,ElvUIParent,TOP,221,-640'
 	E.db.movers.PrivateRaidWarningMover = (scaled and 'TOP,ElvUIParent,TOP,0,-154') or 'TOP,ElvUIParent,TOP,0,-200'
@@ -1849,7 +1904,7 @@ function Private:Setup_ElvUI(layout)
 	E.db.movers.TopCenterContainerMover = 'TOP,ElvUIParent,TOP,0,-67'
 	E.db.movers.TorghastChoiceToggle = 'TOP,ElvUIParent,TOP,0,-598'
 	E.db.movers.UIErrorsFrameMover = 'TOP,ElvUIParent,TOP,0,-117'
-	E.db.movers.VehicleLeaveButton = (scaled and 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,450,413') or 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,706,520'
+	E.db.movers.VehicleLeaveButton = (scaled and 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,543,412') or 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,831,552'
 	E.db.movers.VehicleSeatMover = 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,486,1'
 	E.db.movers.VOICECHAT = 'TOPLEFT,ElvUIParent,TOPLEFT,1,-82'
 
@@ -1859,34 +1914,14 @@ function Private:Setup_ElvUI(layout)
 	if layout == 'main' then
 
 		-- Main Player
-		E.db.unitframe.units.player.customTexts.Luckyone_Power.text_format = Private.isRetail and '[luckyone:power:percent-color]' or '[luckyone:power:percent-color]'
 		E.db.unitframe.units.player.castbar.height = 22
 		E.db.unitframe.units.player.castbar.width = 315
-		E.db.unitframe.units.player.power.autoHide = true
 
-		-- Main TargetTarget
-		E.db.unitframe.units.targettarget.width = 278
-
-		-- Main Party (Classic Only)
-		if Private.isClassic then
-			E.db.unitframe.units.party.customTexts.Luckyone_Power = {
-				attachTextTo = 'Power',
-				enable = true,
-				font = Private.Font,
-				fontOutline = Private.Outline,
-				justifyH = 'CENTER',
-				size = 11,
-				text_format = '[luckyone:power:percent-color<%]',
-				xOffset = 88,
-				yOffset = 20
-			}
-			E.db.unitframe.units.party.power.autoHide = true
-			E.db.unitframe.units.party.power.enable = true
-			E.db.unitframe.units.party.power.height = 4
-			E.db.unitframe.units.party.power.text_format = ''
-		end
+		-- Main Party
+		E.db.unitframe.units.party.customTexts.Luckyone_Name.text_format = (Private.isRetail and '[luckyone:name:short-color-friendly]' or '[luckyone:name:short-classcolor]') .. (not Private.isRetail and '[ ||r- >luckyone:healermana:percent]' or '[ ||r- >luckyone:healermana:percent<%]')
 
 		-- Main Raid1
+		E.db.unitframe.units.raid1.buffIndicator.size = 6
 		E.db.unitframe.units.raid1.height = 40
 		E.db.unitframe.units.raid1.raidicon.attachTo = 'TOPRIGHT'
 		E.db.unitframe.units.raid1.raidicon.attachToObject = 'Health'
@@ -1938,13 +1973,6 @@ function Private:Setup_ElvUI(layout)
 		E.db.unitframe.units.raid3.rdebuffs.stack.yOffset = 0
 		E.db.unitframe.units.raid3.rdebuffs.xOffset = 40
 		E.db.unitframe.units.raid3.rdebuffs.yOffset = 8
-		E.db.unitframe.units.raid3.roleIcon.attachTo = 'Frame'
-		E.db.unitframe.units.raid3.roleIcon.damager = false
-		E.db.unitframe.units.raid3.roleIcon.enable = true
-		E.db.unitframe.units.raid3.roleIcon.position = 'LEFT'
-		E.db.unitframe.units.raid3.roleIcon.size = 12
-		E.db.unitframe.units.raid3.roleIcon.xOffset = 0
-		E.db.unitframe.units.raid3.roleIcon.yOffset = 0
 		E.db.unitframe.units.raid3.width = (scaled and 86) or 96
 
 		-- Main misc
@@ -1964,24 +1992,28 @@ function Private:Setup_ElvUI(layout)
 	elseif layout == 'healing' then
 
 		-- Healing Player
-		E.db.unitframe.units.player.customTexts.Luckyone_Power.text_format = Private.isRetail and '[luckyone:power:percent-nocolor]' or '[curpp< • ][luckyone:power:percent-nocolor<%]'
 		E.db.unitframe.units.player.power.attachTextTo = 'Power'
 		E.db.unitframe.units.player.power.autoHide = false
 		E.db.unitframe.units.player.power.detachedWidth = 260
 		E.db.unitframe.units.player.power.detachFromFrame = true
+		E.db.unitframe.units.player.power.enable = true
 		E.db.unitframe.units.player.power.height = 18
 		E.db.unitframe.units.player.power.position = 'CENTER'
 		E.db.unitframe.units.player.power.powerPrediction = true
+		E.db.unitframe.units.player.power.strataAndLevel.frameStrata = 'MEDIUM'
+		E.db.unitframe.units.player.power.strataAndLevel.useCustomStrata = true
 
-		if Private.isClassic or Private.isTBC then
+		if (Private.isClassic or Private.isTBC) then
 			E.db.unitframe.units.player.power.EnergyManaRegen = true
 		end
 
-		-- Healing TargetTarget
-		E.db.unitframe.units.targettarget.width = 260
+		-- Healing Party
+		E.db.unitframe.units.party.customTexts.Luckyone_Name.text_format = Private.isRetail and '[luckyone:name:short-color-friendly]' or '[luckyone:name:short-classcolor]'
 
 		-- Healing Raid1
 		E.db.unitframe.units.raid1.buffIndicator.size = 14
+		E.db.unitframe.units.raid1.buffs.enable = Private.isRetail
+		E.db.unitframe.units.raid1.debuffs.enable = Private.isRetail
 		E.db.unitframe.units.raid1.height = (scaled and 60) or 70
 		E.db.unitframe.units.raid1.raidicon.attachTo = 'RIGHT'
 		E.db.unitframe.units.raid1.raidicon.size = 14
@@ -1995,16 +2027,13 @@ function Private:Setup_ElvUI(layout)
 		E.db.unitframe.units.raid1.rdebuffs.stack.yOffset = 0
 		E.db.unitframe.units.raid1.rdebuffs.xOffset = 60
 		E.db.unitframe.units.raid1.rdebuffs.yOffset = 27
-		E.db.unitframe.units.raid1.roleIcon.attachTo = 'Frame'
-		E.db.unitframe.units.raid1.roleIcon.damager = false
-		E.db.unitframe.units.raid1.roleIcon.position = 'LEFT'
 		E.db.unitframe.units.raid1.roleIcon.size = 16
-		E.db.unitframe.units.raid1.roleIcon.xOffset = 0
-		E.db.unitframe.units.raid1.roleIcon.yOffset = 0
 		E.db.unitframe.units.raid1.width = 140
 
 		-- Healing Raid2
 		E.db.unitframe.units.raid2.buffIndicator.size = 14
+		E.db.unitframe.units.raid2.buffs.enable = Private.isRetail
+		E.db.unitframe.units.raid2.debuffs.enable = Private.isRetail
 		E.db.unitframe.units.raid2.height = (scaled and 44) or 54
 		E.db.unitframe.units.raid2.raidicon.attachTo = 'RIGHT'
 		E.db.unitframe.units.raid2.raidicon.size = 14
@@ -2018,12 +2047,7 @@ function Private:Setup_ElvUI(layout)
 		E.db.unitframe.units.raid2.rdebuffs.stack.yOffset = 0
 		E.db.unitframe.units.raid2.rdebuffs.xOffset = 60
 		E.db.unitframe.units.raid2.rdebuffs.yOffset = 20
-		E.db.unitframe.units.raid2.roleIcon.attachTo = 'Frame'
-		E.db.unitframe.units.raid2.roleIcon.damager = false
-		E.db.unitframe.units.raid2.roleIcon.position = 'LEFT'
 		E.db.unitframe.units.raid2.roleIcon.size = 16
-		E.db.unitframe.units.raid2.roleIcon.xOffset = 0
-		E.db.unitframe.units.raid2.roleIcon.yOffset = 0
 		E.db.unitframe.units.raid2.width = 140
 
 		-- Healing Raid3
@@ -2041,13 +2065,6 @@ function Private:Setup_ElvUI(layout)
 		E.db.unitframe.units.raid3.rdebuffs.stack.yOffset = 0
 		E.db.unitframe.units.raid3.rdebuffs.xOffset = 60
 		E.db.unitframe.units.raid3.rdebuffs.yOffset = 11
-		E.db.unitframe.units.raid3.roleIcon.attachTo = 'Frame'
-		E.db.unitframe.units.raid3.roleIcon.damager = false
-		E.db.unitframe.units.raid3.roleIcon.enable = true
-		E.db.unitframe.units.raid3.roleIcon.position = 'LEFT'
-		E.db.unitframe.units.raid3.roleIcon.size = 12
-		E.db.unitframe.units.raid3.roleIcon.xOffset = 0
-		E.db.unitframe.units.raid3.roleIcon.yOffset = 0
 		E.db.unitframe.units.raid3.width = 140
 
 		-- Healing misc
@@ -2069,5 +2086,16 @@ function Private:Setup_ElvUI(layout)
 			E.db.actionbar.bar2.mouseover = true
 			E.db.actionbar.bar3.mouseover = true
 		end
+	end
+
+	-- Custom nonRetail changes
+	if Private.itsLuckyone and not Private.isRetail then
+		-- Extra bar next to the left chat panel
+		E.db.actionbar.bar4.enable = true
+		E.db.actionbar.bar4.buttons = 8
+		E.db.actionbar.bar4.buttonsPerRow = 1
+		E.db.actionbar.bar4.buttonSize = 25
+		E.db.actionbar.bar4.point = 'BOTTOMRIGHT'
+		E.db.movers.ElvAB_4 = 'BOTTOMLEFT,ElvUIParent,BOTTOMLEFT,486,2'
 	end
 end