+++ CHANGELOG.md
@@ -1,4 +1,31 @@
-﻿## Version 4.01 [30-01-2026]
+﻿## Version 4.02 [05-02-2026]
+
+**All WoW Versions:**  
+- ElvUI 15.03+ compatibility  
+- Added back a updated version of the player mana bar (Healing Layout)  
+- Added healermana text on party frames for the DPS & Tank layout  
+- Chat installer print will now remind you to create and save an editmode profile  
+- Fixed default position for the bag bar in case someone enables it  
+- Reworked Target Frame (Added TargetTarget text and moved text positions)  
+- Tweaked some more values in the 1080p profiles (Positioning)  
+- Updated BigWigs profile strings (Messages Y-Offset, Bars anchorpoint)  
+- Updated cooldown and duration swipes, edges and blings for most elements  
+- Updated CVars and Nameplate CVars setup buttons and config descriptions  
+- Updated Party Frames (Debuffs anchorpoint, growth, height)  
+
+**Retail:**  
+- Added initial Buffs and Debuffs setup for Raid Frames (Healing Layout)  
+- Enabled the new ElvUI Nameplates option "Lock Friendly NPC Plates" by default  
+- Fixed error in API.lua:65 attempt to compare local 'name' (a secret value)  
+- Reworked Arena Frames (Style will match the new Target Frame)  
+- Reworked Boss Frames (Style will match the new Target Frame)  
+- Updated WindTools profile (Skins)  
+
+**TBC:**  
+- Fixed custom skins loading even when the option was disabled  
+- Fixed power sometimes displaying inf% for people out of range  
+
+## Version 4.01 [30-01-2026]
 
 **All WoW Versions:**  
 - ElvUI 15.02+ compatibility